make delete_prs.py use the new bucket for PRs

+++ s3utils/delete_prs.py
@@ -2,8 +2,11 @@
 import os
 import boto3
 import botocore
+from datetime import datetime, timedelta, timezone
 
-BUCKET = 'status-im'
+BUCKET = 'status-im-prs'
+# remove builds older than 30 days
+CUTOFF = 30
 
 session = boto3.session.Session()
 s3 = session.client('s3',
@@ -19,18 +22,14 @@ def exists(name):
         return False
     return True
 
+threshold = datetime.now(timezone.utc) - timedelta(days=CUTOFF)
+
 for f in s3.list_objects_v2(Bucket=BUCKET, MaxKeys=9999)['Contents']:
     name = f['Key']
-    build_type = name.split('-')[-1].split('.')[0]
-    if build_type == 'e2e':
-        # if an e2e build has a release or a nightly it's not a PR
-        not_pr = any([
-          exists(name.replace('e2e', 'release')),
-          exists(name.replace('e2e', 'nightly'))
-        ])
-        if not_pr:
-            continue
-    elif build_type != 'pr':
+    modified = f['LastModified']
+    if name == 'index.html':
+        continue
+    if modified > threshold:
         continue
     print('DELETING: {}:{}'.format(BUCKET, name))
     s3.delete_object(Bucket=BUCKET, Key=name)
